Update agent.sh
Added installer check for debian that will lead to the unsupported message if on an architecture for which there isn't a suitable amonagent binary. I'm doing this because the installer ran fine on my armv7l machine, but the amonagent binary will not execute on it, nor would amonagent32 (obviously).

Adjusted string comparison expressions to be more robust (if [ $DISTRO == "astring"] is not as good as if [ "$DISTRO" == "astring"] because when DISTRO is undefined, bash will crap its pants.